# docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: 电力交易智能辅助决策系统 API
  description: |
    配套开发基线 5.5 API 目录总表。统一响应封装 ApiResponse；分页用 PageResult；
    鉴权 JWT Bearer；错误码分段见 README 第 4 条。所有路径省略网关前缀 /api/v1。
  version: 1.0.0
  license:
    name: 项目内部使用
    url: https://api.ptidss.example.com/license
  contact:
    name: PTIDSS 架构组
servers:
  - url: https://api.ptidss.example.com/api/v1
    description: 生产（示例域名）
  - url: https://api-test.ptidss.example.com/api/v1
    description: 测试
...
security:
  - bearerAuth: []
paths:
...
  /admin/regions:
    get:
      operationId: getAdminRegions
      summary: 区域列表（多省配置化，评审决议⑤）
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - { name: keyword, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string, enum: [enabled, disabled, pending] } }
      responses:
        '200':
          description: 区域列表
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/ApiResponse' }
                  - properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/SysRegion' }
        '4XX': { $ref: '#/components/responses/BizError' }
    post:
      operationId: postAdminRegions
      summary: 新建区域
      tags: [admin]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SysRegion' }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '4XX': { $ref: '#/components/responses/BizError' }
    put:
      operationId: putAdminRegions
      summary: 更新区域
      tags: [admin]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SysRegion' }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '4XX': { $ref: '#/components/responses/BizError' }
  /admin/regions/{id}:
    get:
      operationId: getAdminRegionsDetail
      summary: 区域详情
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: 区域
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/ApiResponse' }
                  - properties: { data: { $ref: '#/components/schemas/SysRegion' } }
        '4XX': { $ref: '#/components/responses/BizError' }
    delete:
      operationId: deleteAdminRegions
      summary: 删除区域（软删）
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '4XX': { $ref: '#/components/responses/BizError' }
...
  /admin/roles:
    get:
      operationId: getAdminRoles
      summary: 角色列表（固定 7 类编码）
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - { name: keyword, in: query, schema: { type: string } }
      responses:
        '200':
          description: 角色列表
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/ApiResponse' }
                  - properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/SysRole' }
        '4XX': { $ref: '#/components/responses/BizError' }
    post:
      operationId: postAdminRoles
      summary: 新建角色
      tags: [admin]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SysRole' }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '4XX': { $ref: '#/components/responses/BizError' }
    put:
      operationId: putAdminRoles
      summary: 更新角色
      tags: [admin]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SysRole' }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '4XX': { $ref: '#/components/responses/BizError' }
  /admin/roles/{id}:
    get:
      operationId: getAdminRolesDetail
      summary: 角色详情
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: 角色
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/ApiResponse' }
                  - properties: { data: { $ref: '#/components/schemas/SysRole' } }
        '4XX': { $ref: '#/components/responses/BizError' }
    delete:
      operationId: deleteAdminRoles
      summary: 删除角色（软删）
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '4XX': { $ref: '#/components/responses/BizError' }
  /admin/roles/{id}/permissions:
    get:
      operationId: getAdminRolesPermissions
      summary: 角色已授权限
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: 权限 ID 列表
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/ApiResponse' }
                  - properties:
                      data: { type: array, items: { type: integer } }
        '4XX': { $ref: '#/components/responses/BizError' }
    put:
      operationId: putAdminRolesPermissions
      summary: 保存角色权限（全量覆盖）
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                permissionIds: { type: array, items: { type: integer } }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '4XX': { $ref: '#/components/responses/BizError' }
  /admin/permissions:
    get:
      operationId: getAdminPermissions
      summary: 权限列表（三级权限：menu/api/data）
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - { name: keyword, in: query, schema: { type: string } }
        - { name: resourceType, in: query, schema: { type: string, enum: [menu, api, data] } }
      responses:
        '200':
          description: 权限列表
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/ApiResponse' }
                  - properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/SysPermission' }
        '4XX': { $ref: '#/components/responses/BizError' }
    post:
      operationId: postAdminPermissions
      summary: 新建权限
      tags: [admin]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SysPermission' }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '4XX': { $ref: '#/components/responses/BizError' }
    put:
      operationId: putAdminPermissions
      summary: 更新权限
      tags: [admin]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SysPermission' }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '4XX': { $ref: '#/components/responses/BizError' }
  /admin/permissions/{id}:
    delete:
      operationId: deleteAdminPermissions
      summary: 删除权限（软删）
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '4XX': { $ref: '#/components/responses/BizError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BizError:
      description: 业务错误（错误码分段见 README）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
    ApiResponse:
      type: object
      required: [code, message, data, traceId]
      properties:
        code: { type: integer, description: 0 成功；非 0 见错误码分段表, example: 0 }
        message: { type: string, example: success }
        data: { type: object, nullable: true, description: 业务数据（随接口变化） }
        traceId: { type: string, description: 全链路追踪 ID }
    ErrorResponse:
      type: object
      properties:
        code: { type: integer, example: 14001 }
        message: { type: string }
        traceId: { type: string }
        detail: { type: object, nullable: true, description: 参数错误明细 }
...
    SysRegion:
      type: object
      properties:
        id: { type: string, description: 雪花 ID（JSON 序列化为字符串） }
        regionCode: { type: string }
        regionName: { type: string }
        status: { type: string, enum: [enabled, disabled, pending] }
        launchOrder: { type: integer }
...
    SysRole:
      type: object
      description: 角色（编码固定 7 类，与 PRD 七类角色对齐；新建角色受服务端 DDL CHECK 约束，仅可更新 description/status）
      properties:
        id: { type: string, description: 雪花 ID（JSON 序列化为字符串） }
        roleName: { type: string }
        roleCode:
          type: string
          enum: [trader, analyst, settlement, admin, manager, compliance, mobile]
        description: { type: string }
        status: { type: string }
    SysPermission:
      type: object
      properties:
        id: { type: string, description: 雪花 ID（JSON 序列化为字符串） }
        permName: { type: string }
        permCode: { type: string }
        resourceType: { type: string, enum: [menu, api, data] }
        resourcePattern: { type: string }
...